requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
jinja2
python-multipart
matplotlib
rapidfuzz
unidecode
openpyxl
pytest
python-docx
